finished

+++ main.cpp
@@ -20,7 +20,7 @@ int main(int argc, char* argv[]) {
     my_vector_t<int> vec;
     std::cout << vec.size() << std::endl;
     for (size_t i = 0; i < 5; ++i) {
-        vec.push_back(i);
+        vec.push_back(static_cast<int>(i));
     }
 
     for (size_t i = 0; i < vec.size(); ++i) {
